Moved ev-registry and did classes under flex-market
This is to make is clearer that they are in the flex-market domain
